refactor: 예약 상태 목록에 대한 Controller 제거
- 불필요한 Controller 분리 제거
- Enum 목록을 Controller에서 바로 반환하도록 변경
