+++ src/main/java/roomescape/reservation/ui/ReservationRestController.java
@@ -4,6 +4,7 @@ import static roomescape.auth.domain.AuthRole.ADMIN;
 import static roomescape.auth.domain.AuthRole.MEMBER;
 
 import jakarta.validation.Valid;
+import java.util.Arrays;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -18,10 +19,12 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.domain.MemberAuthInfo;
 import roomescape.auth.domain.RequiresRole;
 import roomescape.reservation.application.ReservationService;
+import roomescape.reservation.domain.BookingState;
 import roomescape.reservation.ui.dto.request.AvailableReservationTimeRequest;
 import roomescape.reservation.ui.dto.request.CreateReservationRequest;
 import roomescape.reservation.ui.dto.request.ReservationsByfilterRequest;
 import roomescape.reservation.ui.dto.response.AvailableReservationTimeResponse;
+import roomescape.reservation.ui.dto.response.BookingStateResponse;
 import roomescape.reservation.ui.dto.response.ReservationResponse;
 
 @RestController
@@ -93,4 +96,14 @@ public class ReservationRestController {
 
         return ResponseEntity.ok(availableReservationTimes);
     }
+
+    @GetMapping("/reservations/states")
+    @RequiresRole(authRoles = {ADMIN, MEMBER})
+    public ResponseEntity<List<BookingStateResponse>> getBookingStateOptions() {
+        return ResponseEntity.ok()
+                .body(Arrays.stream(BookingState.values())
+                        .map(BookingStateResponse::from)
+                        .toList());
+    }
+
 }
