--- src/main/java/roomescape/reservation/ui/ReservationStatusRestController.java
@@ -1,11 +1,0 @@
-package roomescape.reservation.ui;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class ReservationStatusRestController {
-
-
-}
